+++ src/main/java/camp/nextstep/edu/kitchenpos/ordertable/dao/OrderTableDao.java
@@ -1,6 +1,6 @@
 package camp.nextstep.edu.kitchenpos.ordertable.dao;
 
-import camp.nextstep.edu.kitchenpos.ordertable.model.OrderTable;
+import camp.nextstep.edu.kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
